Reorganise drawing tool view components Move drawing tool views to plugins/drawingTools/components
