+++ packages/lib-classifier/src/plugins/drawingTools/components/DeleteButton.js
