+++ packages/lib-classifier/src/plugins/drawingTools/components/DragHandle/DragHandle.js
