+++ packages/lib-classifier/src/plugins/drawingTools/components/DragHandle/index.js
