+++ packages/lib-classifier/src/plugins/drawingTools/components/DrawingToolRoot.spec.js
