+++ packages/lib-classifier/src/plugins/drawingTools/components/Line.spec.js
