+++ packages/lib-classifier/src/plugins/drawingTools/components/Point.js
