+++ packages/lib-classifier/src/plugins/drawingTools/components/Point.spec.js
