+++ packages/lib-classifier/src/plugins/drawingTools/components/draggable/draggable.js
