+++ packages/lib-classifier/src/plugins/drawingTools/components/draggable/draggable.stories.js
