+++ packages/lib-classifier/src/plugins/drawingTools/components/draggable/index.js
