+++ packages/lib-classifier/src/plugins/drawingTools/components/index.js
@@ -1,0 +1,6 @@
+export { default as draggable } from './draggable'
+export { default as DragHandle } from './DragHandle'
+export { default as DeleteButton } from './DeleteButton'
+export { default as DrawingToolRoot } from './DrawingToolRoot'
+export { default as Line } from './Line'
+export { default as Point } from './Point'
