--- packages/lib-classifier/src/plugins/tasks/DrawingTask/components/tools/index.js
@@ -1,2 +1,0 @@
-export { default as Line } from './Line'
-export { default as Point } from './Point'
